# apps/python/app/models/nestjs_models.py
# generated by datamodel-codegen:
#   filename:  openapi.json
#   timestamp: 2025-03-16T19:51:57+00:00

from __future__ import annotations

from pydantic import BaseModel


class Object(BaseModel):
    pass


class CompletionRequestDto(BaseModel):
    pass