# apps/python/app/utils/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "description": "Authenticate a user with email and password",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "email": "user@example.com",
                  "password": "password123"
                },
                "$ref": "#/components/schemas/Object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "user": {
                        "id": "1",
                        "email": "user@example.com",
                        "name": "John Doe",
                        "role": "user"
                      }
                    },
                    "status": 200,
                    "message": "Login successful"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "User login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "description": "Register a new user account",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "email": "newuser@example.com",
                  "password": "password123",
                  "name": "New User"
                },
                "$ref": "#/components/schemas/Object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration successful",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "user": {
                        "id": "2",
                        "email": "newuser@example.com",
                        "name": "New User",
                        "role": "user"
                      }
                    },
                    "status": 201,
                    "message": "Registration successful"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or email already exists"
          }
        },
        "summary": "User registration",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "description": "Log out the current user and clear session",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": null,
                    "status": 200,
                    "message": "Logout successful"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "User logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/profile": {
      "get": {
        "description": "Get the current user's profile information",
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "id": "1",
                      "email": "user@example.com",
                      "name": "John Doe",
                      "role": "user",
                      "createdAt": "2023-01-01T00:00:00.000Z",
                      "updatedAt": "2023-01-01T00:00:00.000Z",
                      "emailVerified": true
                    },
                    "status": 200,
                    "message": "Profile retrieved successfully"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": null,
                    "status": 401,
                    "message": "Authentication required",
                    "error": {
                      "code": "UNAUTHORIZED",
                      "message": "Authentication required"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get user profile",
        "tags": [
          "auth"
        ]
      },
      "patch": {
        "description": "Update the current user's profile information",
        "operationId": "AuthController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "name": "Updated Name",
                  "email": "updated.email@example.com"
                },
                "$ref": "#/components/schemas/Object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "id": "1",
                      "email": "updated.email@example.com",
                      "name": "Updated Name",
                      "createdAt": "2023-01-01T00:00:00.000Z",
                      "updatedAt": "2023-01-02T00:00:00.000Z"
                    },
                    "status": 200,
                    "message": "Profile updated successfully"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or email already in use"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Update user profile",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/change-password": {
      "post": {
        "description": "Change the current user's password",
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "currentPassword": "oldPassword123",
                  "newPassword": "newPassword123",
                  "confirmPassword": "newPassword123"
                },
                "$ref": "#/components/schemas/Object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": null,
                    "status": 200,
                    "message": "Password changed successfully"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Current password is incorrect or passwords don't match"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Change password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/account": {
      "delete": {
        "description": "Permanently delete the current user's account",
        "operationId": "AuthController_deleteAccount",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "password": "password123",
                  "confirmation": "DELETE"
                },
                "$ref": "#/components/schemas/Object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": null,
                    "status": 200,
                    "message": "Account deleted successfully"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Password is incorrect or confirmation is invalid"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Delete account",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "description": "Get a new access token using the refresh token cookie",
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": null,
                    "status": 200,
                    "message": "Token refreshed successfully"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Refresh access token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/admin-profile": {
      "get": {
        "description": "Retrieve the admin profile information (admin only)",
        "operationId": "AuthController_getAdminProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Admin profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "id": "1",
                      "email": "admin@example.com",
                      "name": "Admin User",
                      "role": "admin",
                      "createdAt": "2023-01-01T00:00:00.000Z"
                    },
                    "status": 200,
                    "message": "Admin profile retrieved successfully"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get admin profile",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/verify-email": {
      "get": {
        "description": "Verify user email with token",
        "operationId": "AuthController_verifyEmail",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "description": "Email verification token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "verified": true
                    },
                    "status": 200,
                    "message": "Email verified successfully"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired token"
          }
        },
        "summary": "Verify email",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/resend-verification": {
      "post": {
        "description": "Resend the email verification link",
        "operationId": "AuthController_resendVerification",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Verification email sent",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": null,
                    "status": 200,
                    "message": "Verification email sent"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Email already verified"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Resend verification email",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/email-status": {
      "get": {
        "description": "Check if the user's email is verified",
        "operationId": "AuthController_getEmailStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Email verification status",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "verified": true
                    },
                    "status": 200,
                    "message": "Email verification status"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get email verification status",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Fullstack Boilerplate API",
    "description": "The API documentation for the Fullstack Boilerplate project",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "cookie": {
        "type": "apiKey",
        "in": "Cookie",
        "name": "connect.sid"
      }
    },
    "schemas": {
      "Object": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}